require 專案的模組改用 require.main.require ，包裝 json-api, data-admin
