+++ lib/env-setting.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var reloadify = require('./reloadify');
-var bootlint = require('./bootlint');
+var reloadify = require.main.require('./lib/reloadify');
+var bootlint = require.main.require('./lib/bootlint');
 
 var enviromentSetting = function (app) {
     switch (app.get('env')) {
