+++ models/members.js
@@ -2,7 +2,7 @@
 
 var mongoose = require('mongoose');
 
-var dbManager = require('../lib/db-manager.js');
+var dbManager = require.main.require('./lib/db-manager.js');
 
 var memberSchema = new mongoose.Schema({
     name: String,
