+++ timesheet.js
@@ -6,8 +6,8 @@ var path = require('path');
 var express = require('express');
 
 //load inner modules
-var enviromentSetting = require('./lib/env-setting.js');
-var routes = require('./routes/routes.js');
+var enviromentSetting = require.main.require('./lib/env-setting.js');
+var routes = require.main.require('./routes/routes.js');
 
 var app = express();
